Fix #51
