+++ src/parser/mod.rs
@@ -1,6 +1,5 @@
 pub mod cumin;
 pub mod expr;
-pub mod logic;
 pub mod module;
 pub mod statement;
 pub mod typing;
